+++ tests/utilities.py
@@ -6,7 +6,6 @@ from datetime import datetime
 from harmony.util import bbox_to_geometry
 from pystac import Asset, Catalog, Item
 
-
 Granule = namedtuple('Granule', ['url', 'media_type', 'roles'])