+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.0.2
+### 2024-04-05
+
+This version of HyBIG correctly handles missing/bad input data marked by _FillValue or NoData.
+Anytime a bad value occurs in the input raster, the output png image will set to transparent.
+
 ## v1.0.1
 ### 2024-04-05